+++ packages/sandbox/CHANGELOG.md
@@ -1,5 +1,12 @@
 # sandbox
 
+## 3.0.0-beta.18
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/sandbox@2.0.0-beta.16
+
 ## 3.0.0-beta.17
 
 ### Minor Changes